A single variable is stored on the stack!
Lots of panics internally if there is a completely reasonable error.
If the lhs is not assignable, panic. Unresolved identifier, panic.
There is currently no semantic analysis of the compiler does, as long as
it parses correctly, it will generate code, or panic. I can't put the
errors in the parser because I want the parser to report parse errors. I
also can't put the errors in the code generator because i don't want
that to fail (linking can fail but that is separate.

I think since C is built for single pass compilation that I could
probably have some sort of semantic analysis done during parsing, but
that these errors will be reported after parse errors. This might be
easier and probably faster. I could also have a static analysis layer in
the middle, since I will probably want that anyway once we get some sort
of SSA IR involved.